refactor(observer): split snapshot graph responsibilities
