+++ apps/observer/src/internal.ts
@@ -25,7 +25,10 @@ export * from "./persistence/recoveryArchiveReader.js";
 export * from "./providers/healthCache.js";
 export * from "./providers/registry.js";
 export * from "./reconcile/core.js";
-export * from "./reconcile/graph.js";
+export * from "./reconcile/graph/authoritativeLaunch.js";
+export * from "./reconcile/graph/build.js";
+export * from "./reconcile/graph/evidence.js";
+export * from "./reconcile/graph/providerHealth.js";
 export * from "./reconcile/statusProjection.js";
 export * from "./runtime/api.js";
 export * from "./runtime/eventBus.js";
